gen: Refactor generator
Refactor `topology/generator.py` in multiple files.
