+++ python/topology/common.py
@@ -1,4 +1,4 @@
-# Copyright 2018 ETH Zurich
+# Copyright 2018 ETH Zurich, Anapaya Systems
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,6 +11,45 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+# Stdlib
+import os
+
+# SCION
+from lib.packet.scion_addr import ISD_AS
+
+COMMON_DIR = 'endhost'
+
+SCION_SERVICE_NAMES = (
+    "BeaconService",
+    "CertificateService",
+    "BorderRouters",
+    "PathService",
+    "DiscoveryService",
+)
+
+
+class TopoID(ISD_AS):
+    def ISD(self):
+        return "ISD%s" % self.isd_str()
+
+    def AS(self):
+        return "AS%s" % self.as_str()
+
+    def AS_file(self):
+        return "AS%s" % self.as_file_fmt()
+
+    def file_fmt(self):
+        return "%s-%s" % (self.isd_str(), self.as_file_fmt())
+
+    def base_dir(self, out_dir):
+        return os.path.join(out_dir, self.ISD(), self.AS_file())
+
+    def __lt__(self, other):
+        return str(self) < str(other)
+
+    def __repr__(self):
+        return "<TopoID: %s>" % self
 
 
 def _prom_addr_br(br_ele):
@@ -30,3 +69,21 @@ def _get_pub(topo_addr):
     if pub is not None:
         return pub
     return topo_addr['IPv4']
+
+
+def _get_pub_ip(topo_addr):
+        return _get_pub(topo_addr)["Public"]["Addr"].ip
+
+
+def _get_l4_port(topo_addr):
+    return _get_pub(topo_addr)["Public"]["L4Port"]
+
+
+def _srv_iter(topo_dicts, out_dir, common=False):
+    for topo_id, as_topo in topo_dicts.items():
+        base = topo_id.base_dir(out_dir)
+        for service in SCION_SERVICE_NAMES:
+            for elem in as_topo[service]:
+                yield topo_id, as_topo, os.path.join(base, elem)
+        if common:
+            yield topo_id, as_topo, os.path.join(base, COMMON_DIR)
